Factor out EventManager
